# schemas/logweir-api-v1.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Logweir product API",
    "version": "1.0.0-alpha.1",
    "description": "The bounded shared-console product API (PLAT-17.1). Not a Kubernetes proxy: no generic resource path, no Secret, Pod, log, exec or Job access, no delete, and no arbitrary patch. Every error is application/problem+json. Unsafe methods require Content-Type: application/json and an Origin equal to the configured public origin. Durable creates require Idempotency-Key. Capabilities without a route are advertised false in /api/v1/session."
  },
  "paths": {
...
    "/api/v1/namespaces/{ns}/backups": {
      "get": {
        "operationId": "listBackups",
        "summary": "Backup projections with a status summary.",
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "description": "An explicitly granted namespace. An ungranted namespace is namespace_forbidden whether or not it exists.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, 1 to 200. Default 50.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "The opaque nextCursor of the previous page. Bound to actor, route, namespace and filters; expires after 15 minutes.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelSelector",
            "in": "query",
            "required": false,
            "description": "Equality-only label selector, key=value terms separated by commas.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupList"
                }
              }
            }
          },
          "default": {
            "description": "An application/problem+json error; the possible codes are listed in x-logweir-problem-codes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-logweir-problem-codes": [
          "cursor_expired",
          "cursor_invalid",
          "forbidden",
          "header_not_allowed",
          "internal_error",
          "kubernetes_unavailable",
          "malformed_request",
          "method_not_allowed",
          "misdirected_request",
          "namespace_forbidden",
          "rate_limited",
          "session_expired",
          "unauthenticated",
          "upstream_timeout",
          "validation_failed"
        ]
      }
    },
...
    "/api/v1/namespaces/{ns}/approvals": {
      "get": {
        "operationId": "listApprovals",
        "summary": "Approval metadata; no documents.",
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "description": "An explicitly granted namespace. An ungranted namespace is namespace_forbidden whether or not it exists.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size, 1 to 200. Default 50.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "The opaque nextCursor of the previous page. Bound to actor, route, namespace and filters; expires after 15 minutes.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "labelSelector",
            "in": "query",
            "required": false,
            "description": "Equality-only label selector, key=value terms separated by commas.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalList"
                }
              }
            }
          },
          "default": {
            "description": "An application/problem+json error; the possible codes are listed in x-logweir-problem-codes.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-logweir-problem-codes": [
          "cursor_expired",
          "cursor_invalid",
          "forbidden",
          "header_not_allowed",
          "internal_error",
          "kubernetes_unavailable",
          "malformed_request",
          "method_not_allowed",
          "misdirected_request",
          "namespace_forbidden",
          "rate_limited",
          "session_expired",
          "unauthenticated",
          "upstream_timeout",
          "validation_failed"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Approval": {
        "description": "Public approval metadata. The documents themselves are only on the packet route.",
        "type": "object",
        "required": [
          "approvalBytesLength",
          "conditions",
          "name",
          "namespace",
          "planHash",
          "resourceVersion",
          "sidecarBytesLength",
          "subjectRef",
          "uid"
        ],
        "properties": {
          "approvalBytesLength": {
            "description": "The approval document's length in bytes.",
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "approver": {
            "description": "The approver subject from the roster.",
            "type": "string",
            "nullable": true
          },
          "conditions": {
            "description": "The status conditions, at most 16.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConditionView"
            }
          },
          "createdAt": {
            "description": "When the object was created.",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "matchedKeyId": {
            "description": "The approver key that verified it.",
            "type": "string",
            "nullable": true
          },
          "name": {
            "description": "The object name.",
            "type": "string"
          },
          "namespace": {
            "description": "The namespace.",
            "type": "string"
          },
          "planHash": {
            "description": "The plan hash the create form supplied.",
            "type": "string"
          },
          "resourceVersion": {
            "description": "The resourceVersion.",
            "type": "string"
          },
          "selfAttestedRisk": {
            "description": "Whether approver and requester may be the same principal.",
            "type": "boolean",
            "nullable": true
          },
          "sidecarBytesLength": {
            "description": "The sidecar's length in bytes.",
            "type": "integer",
            "format": "uint",
            "minimum": 0.0
          },
          "subjectRef": {
            "description": "The subject.",
            "$ref": "#/components/schemas/SubjectRefView"
          },
          "ticket": {
            "description": "The change ticket the document names.",
            "type": "string",
            "nullable": true
          },
          "uid": {
            "description": "The Kubernetes UID.",
            "type": "string"
          },
          "verified": {
            "description": "Whether the controller verified the approval.",
            "type": "boolean",
            "nullable": true
          },
          "verifiedSubject": {
            "description": "The exact verified subject.",
            "$ref": "#/components/schemas/VerifiedSubjectView",
            "nullable": true
          }
        }
      },
      "ApprovalList": {
        "type": "object",
        "required": [
          "items",
          "page",
          "requestId"
        ],
        "properties": {
          "items": {
            "description": "The items on this page.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approval"
            }
          },
          "page": {
            "description": "Paging.",
            "$ref": "#/components/schemas/Page"
          },
          "requestId": {
            "description": "The request ID.",
            "type": "string"
          }
        }
      },
...
      "ArchiveView": {
        "description": "An archive location and the name of the Secret that reaches it.",
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "credentialRef": {
            "description": "The name of the Secret carrying the object-store credential. A name only; this service never reads Secrets.",
            "$ref": "#/components/schemas/NameRef",
            "nullable": true
          },
          "url": {
            "description": "The object-store URL, with any userinfo redacted.",
            "type": "string"
          }
        }
      },
...
      "Backup": {
        "description": "A backup run.",
        "type": "object",
        "required": [
          "archive",
          "deadlineSeconds",
          "name",
          "namespace",
          "operation",
          "resourceVersion",
          "sourceRef",
          "topics",
          "triggeredBy",
          "uid"
        ],
        "properties": {
          "archive": {
            "description": "Where the archive is written.",
            "$ref": "#/components/schemas/ArchiveView"
          },
          "backupId": {
            "description": "The archive backup ID.",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "description": "When the object was created.",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deadlineSeconds": {
            "description": "The Job deadline.",
            "type": "integer",
            "format": "int64"
          },
          "manifestKey": {
            "description": "The manifest key.",
            "type": "string",
            "nullable": true
          },
          "name": {
            "description": "The object name.",
            "type": "string"
          },
          "namespace": {
            "description": "The namespace.",
            "type": "string"
          },
          "observedAuth": {
            "description": "The identity the run presented: mode and username, never a password.",
            "$ref": "#/components/schemas/ObservedAuthView",
            "nullable": true
          },
          "operation": {
            "description": "The normalized status summary.",
            "$ref": "#/components/schemas/OperationSummary"
          },
          "records": {
            "description": "Records archived.",
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "resourceVersion": {
            "description": "The resourceVersion.",
            "type": "string"
          },
          "schedule": {
            "description": "The schedule that created the run.",
            "type": "string",
            "nullable": true
          },
          "slot": {
            "description": "The schedule slot.",
            "type": "string",
            "nullable": true
          },
          "sourceRef": {
            "description": "The source connection.",
            "$ref": "#/components/schemas/NameRef"
          },
          "topics": {
            "description": "Named topics.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "triggeredBy": {
            "description": "`schedule` or `manual`.",
            "type": "string"
          },
          "uid": {
            "description": "The Kubernetes UID.",
            "type": "string"
          },
          "windowCovered": {
            "description": "The covered window, epoch milliseconds, end exclusive.",
            "$ref": "#/components/schemas/WindowCoveredView",
            "nullable": true
          }
        }
      },
      "BackupList": {
        "type": "object",
        "required": [
          "items",
          "page",
          "requestId"
        ],
        "properties": {
          "items": {
            "description": "The items on this page.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Backup"
            }
          },
          "page": {
            "description": "Paging.",
            "$ref": "#/components/schemas/Page"
          },
          "requestId": {
            "description": "The request ID.",
            "type": "string"
          }
        }
      },
...
      "ConditionView": {
        "description": "A status condition, with a bounded message.",
        "type": "object",
        "required": [
          "status",
          "type"
        ],
        "properties": {
          "lastTransitionTime": {
            "description": "When the condition last changed status or reason.",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "message": {
            "description": "The message, at most 1024 bytes.",
            "type": "string",
            "nullable": true
          },
          "reason": {
            "description": "The CamelCase reason.",
            "type": "string",
            "nullable": true
          },
          "status": {
            "description": "`True`, `False` or `Unknown`.",
            "type": "string"
          },
          "type": {
            "description": "The condition type.",
            "type": "string"
          }
        }
      },
...
      "FieldError": {
        "description": "One field-level validation failure.",
        "type": "object",
        "required": [
          "code",
          "field",
          "message"
        ],
        "properties": {
          "code": {
            "description": "A stable machine-readable code for this field, e.g. `required`.",
            "type": "string"
          },
          "field": {
            "description": "The JSON path of the field, e.g. `topics[2]`, or a header or query parameter name.",
            "type": "string"
          },
          "message": {
            "description": "A human-readable sentence. Never echoes a credential.",
            "type": "string"
          }
        }
      },
...
      "NameRef": {
        "description": "A same-namespace reference by name.",
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "description": "The referenced object's name, in the same namespace.",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ObservedAuthView": {
        "description": "The identity a run presented.",
        "type": "object",
        "properties": {
          "mode": {
            "description": "`plaintext` or `scramSha512`.",
            "type": "string",
            "nullable": true
          },
          "username": {
            "description": "The SASL principal presented.",
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "OperationState": {
        "description": "The product lifecycle state.",
        "oneOf": [
          {
            "description": "Accepted and waiting: not yet observed by the controller, or held for an approval.",
            "type": "string",
            "enum": [
              "pending"
            ]
          },
          {
            "description": "Waiting for execution capacity. Not produced by current resources.",
            "type": "string",
            "enum": [
              "queued"
            ]
          },
          {
            "description": "The controller is materializing execution inputs.",
            "type": "string",
            "enum": [
              "preparing"
            ]
          },
          {
            "description": "The runner Job exists and has not finished.",
            "type": "string",
            "enum": [
              "running"
            ]
          },
          {
            "description": "The run finished and its signed evidence has not been verified yet.",
            "type": "string",
            "enum": [
              "verifying"
            ]
          },
          {
            "description": "The run exited 0.",
            "type": "string",
            "enum": [
              "succeeded"
            ]
          },
          {
            "description": "The run finished without a pass, crashed, or failed operationally.",
            "type": "string",
            "enum": [
              "failed"
            ]
          },
          {
            "description": "A guard or the controller refused the run before anything executed.",
            "type": "string",
            "enum": [
              "refused"
            ]
          },
          {
            "description": "Cancelled. Not produced by current resources (no cancel exists).",
            "type": "string",
            "enum": [
              "cancelled"
            ]
          },
          {
            "description": "A phase this build does not recognise.",
            "type": "string",
            "enum": [
              "unknown"
            ]
          }
        ]
      },
      "OperationSummary": {
        "description": "The status summary carried by list items.",
        "type": "object",
        "required": [
          "state",
          "terminal",
          "verificationState",
          "verifiedSuccess"
        ],
        "properties": {
          "state": {
            "description": "The lifecycle state.",
            "$ref": "#/components/schemas/OperationState"
          },
          "stateReason": {
            "description": "The CamelCase reason.",
            "type": "string",
            "nullable": true
          },
          "terminal": {
            "description": "Whether the state is final.",
            "type": "boolean"
          },
          "verificationState": {
            "description": "The verification state.",
            "$ref": "#/components/schemas/VerificationState"
          },
          "verifiedSuccess": {
            "description": "Result pass AND verification `Valid`.",
            "type": "boolean"
          }
        }
      },
...
      "Page": {
        "description": "One page of a list.",
        "type": "object",
        "required": [
          "limit"
        ],
        "properties": {
          "limit": {
            "description": "The page size that was applied.",
            "type": "integer",
            "format": "uint32",
            "minimum": 0.0
          },
          "nextCursor": {
            "description": "An opaque cursor for the next page, valid for fifteen minutes for the same actor, route, namespace and filters; absent on the last page.",
            "type": "string",
            "nullable": true
          },
          "snapshot": {
            "description": "The Kubernetes list resourceVersion this page was read at.",
            "type": "string",
            "nullable": true
          }
        }
      },
      "Problem": {
        "description": "The rendered problem document.",
        "type": "object",
        "required": [
          "code",
          "detail",
          "requestId",
          "retryable",
          "status",
          "title",
          "type"
        ],
        "properties": {
          "code": {
            "description": "The stable code.",
            "$ref": "#/components/schemas/ProblemCode"
          },
          "detail": {
            "description": "A sentence this service wrote. Never a Kubernetes message.",
            "type": "string"
          },
          "errors": {
            "description": "Field-level failures, when the code is `validation_failed`.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "requestId": {
            "description": "The request ID, also returned as `X-Request-ID`.",
            "type": "string"
          },
          "retryable": {
            "description": "Whether repeating the identical request later can succeed.",
            "type": "boolean"
          },
          "status": {
            "description": "The HTTP status, repeated.",
            "type": "integer",
            "format": "uint16",
            "minimum": 0.0
          },
          "title": {
            "description": "A fixed title for the code.",
            "type": "string"
          },
          "type": {
            "description": "`https://logweir.dev/problems/<code>`.",
            "type": "string"
          }
        }
      },
      "ProblemCode": {
        "description": "The stable problem codes. Clients branch on `code`, never on `detail`.\n\nThe decision's baseline set is here verbatim (`unauthenticated`, `session_expired`, `forbidden`, `namespace_forbidden`, `not_found`, `idempotency_conflict`, `state_conflict`, `approval_required`, `policy_mismatch`, `precondition_failed`, `validation_failed`, `rate_limited`, `kubernetes_unavailable`, `upstream_timeout`, plus the cursor pair). The remainder name transport-level refusals the baseline did not enumerate; each is documented in the OpenAPI document.",
        "oneOf": [
          {
            "description": "400 — the body is not JSON, a query string is malformed or carries an unknown or repeated parameter.",
            "type": "string",
            "enum": [
              "malformed_request"
            ]
          },
          {
            "description": "400 — a header this service refuses outright (`Impersonate-*`).",
            "type": "string",
            "enum": [
              "header_not_allowed"
            ]
          },
          {
            "description": "400 — a durable POST arrived without `Idempotency-Key`.",
            "type": "string",
            "enum": [
              "idempotency_key_required"
            ]
          },
          {
            "description": "400 — `Idempotency-Key` is not 8–128 visible ASCII characters, or is present on a route that does not accept it.",
            "type": "string",
            "enum": [
              "idempotency_key_invalid"
            ]
          },
          {
            "description": "400 — a list cursor failed authentication or names another scope.",
            "type": "string",
            "enum": [
              "cursor_invalid"
            ]
          },
          {
            "description": "401 — no authenticated actor (PLAT-17.2 authenticators).",
            "type": "string",
            "enum": [
              "unauthenticated"
            ]
          },
          {
            "description": "401 — the session is past its signed expiry (PLAT-17.2).",
            "type": "string",
            "enum": [
              "session_expired"
            ]
          },
          {
            "description": "403 — the actor is not permitted this action.",
            "type": "string",
            "enum": [
              "forbidden"
            ]
          },
          {
            "description": "403 — the namespace is not explicitly granted to the actor.",
            "type": "string",
            "enum": [
              "namespace_forbidden"
            ]
          },
          {
            "description": "403 — an unsafe method without an `Origin` exactly equal to the configured public origin.",
            "type": "string",
            "enum": [
              "origin_mismatch"
            ]
          },
          {
            "description": "404 — no such route or no such object.",
            "type": "string",
            "enum": [
              "not_found"
            ]
          },
          {
            "description": "405 — the route exists and does not accept this method.",
            "type": "string",
            "enum": [
              "method_not_allowed"
            ]
          },
          {
            "description": "409 — the idempotency key was already used with a different request.",
            "type": "string",
            "enum": [
              "idempotency_conflict"
            ]
          },
          {
            "description": "409 — the deterministic object exists but was not created by this idempotency scope; it is never adopted.",
            "type": "string",
            "enum": [
              "state_conflict"
            ]
          },
          {
            "description": "409 — an approval is required first (PLAT-19.2).",
            "type": "string",
            "enum": [
              "approval_required"
            ]
          },
          {
            "description": "409 — the bound approval policy does not match (PLAT-19.2).",
            "type": "string",
            "enum": [
              "policy_mismatch"
            ]
          },
          {
            "description": "410 — the list cursor expired or Kubernetes compacted its continue token; restart the list without a cursor.",
            "type": "string",
            "enum": [
              "cursor_expired"
            ]
          },
          {
            "description": "412 — `expectedResourceVersion` is not the object's current version.",
            "type": "string",
            "enum": [
              "precondition_failed"
            ]
          },
          {
            "description": "413 — the request body exceeds its limit.",
            "type": "string",
            "enum": [
              "payload_too_large"
            ]
          },
          {
            "description": "415 — an unsafe method without `Content-Type: application/json`.",
            "type": "string",
            "enum": [
              "unsupported_media_type"
            ]
          },
          {
            "description": "421 — the `Host` header names an authority this listener does not serve (DNS-rebinding defence).",
            "type": "string",
            "enum": [
              "misdirected_request"
            ]
          },
          {
            "description": "422 — the request is well-formed JSON and fails validation.",
            "type": "string",
            "enum": [
              "validation_failed"
            ]
          },
          {
            "description": "429 — Kubernetes throttled the request; see `Retry-After`.",
            "type": "string",
            "enum": [
              "rate_limited"
            ]
          },
          {
            "description": "500 — an invariant this service holds did not hold.",
            "type": "string",
            "enum": [
              "internal_error"
            ]
          },
          {
            "description": "503 — Kubernetes could not be reached or refused this service's own identity.",
            "type": "string",
            "enum": [
              "kubernetes_unavailable"
            ]
          },
          {
            "description": "504 — the Kubernetes call exceeded its 10-second deadline.",
            "type": "string",
            "enum": [
              "upstream_timeout"
            ]
          }
        ]
      },
...
      "SubjectRefView": {
        "description": "What an approval is about.",
        "type": "object",
        "required": [
          "kind",
          "name"
        ],
        "properties": {
          "kind": {
            "description": "`Restore` or `Backup`.",
            "type": "string"
          },
          "name": {
            "description": "The subject name.",
            "type": "string"
          }
        }
      },
...
      "VerifiedSubjectView": {
        "description": "The exact object whose bytes the controller verified.",
        "type": "object",
        "required": [
          "kind",
          "name",
          "namespace",
          "uid"
        ],
        "properties": {
          "kind": {
            "description": "The subject kind.",
            "type": "string"
          },
          "name": {
            "description": "The subject name.",
            "type": "string"
          },
          "namespace": {
            "description": "The subject namespace.",
            "type": "string"
          },
          "uid": {
            "description": "The subject UID.",
            "type": "string"
          }
        }
      },
      "WindowCoveredView": {
        "description": "The covered window.",
        "type": "object",
        "required": [
          "fromMs",
          "toMs"
        ],
        "properties": {
          "fromMs": {
            "description": "Inclusive start, epoch milliseconds.",
            "type": "integer",
            "format": "int64"
          },
          "toMs": {
            "description": "Exclusive end, epoch milliseconds.",
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  },
...
}